+++ initialize.py
@@ -26,10 +26,10 @@
 import sys
 import numpy as np
 from scipy.interpolate import interp1d
-from utils import *
-from PSfunctions import *
-from USStandardAtmosphere import *
-from coordinate import *
+from lib.utils import *
+from lib.PSfunctions import *
+from lib.USStandardAtmosphere import *
+from lib.coordinate import *
 from tools.plot_output import display_6DoF
 from output_result import output_result
