+++ lib/con_aero.py
@@ -27,7 +27,7 @@
 # constraints about aerodynamic conditions
 
 import numpy as np
-from utils import (
+from .utils import (
     dynamic_pressure_pa,
     angle_of_attack_all_rad,
     q_alpha_pa_rad,
@@ -35,7 +35,7 @@ from utils import (
     angle_of_attack_all_array_rad,
     q_alpha_array_pa_rad,
 )
-from coordinate import *
+from .coordinate import *
 
 
 def dynamic_pressure_dimless(pos_eci_e, vel_eci_e, t_e, wind, units):
