+++ lib/con_dynamics.py
@@ -28,7 +28,7 @@
 
 from itertools import chain, repeat
 import numpy as np
-from tools.dynamics import dynamics_velocity, dynamics_velocity_NoAir, dynamics_quaternion
+from .dynamics import dynamics_velocity, dynamics_velocity_NoAir, dynamics_quaternion
 
 
 def equality_dynamics_mass(xdict, pdict, unitdict, condition):
