+++ lib/con_init_terminal_knot.py
@@ -25,7 +25,7 @@
 
 import numpy as np
 from math import cos, radians
-from coordinate import (
+from .coordinate import (
     angular_momentum,
     orbit_energy,
     inclination_rad,
