+++ lib/utils.py
@@ -27,8 +27,8 @@ from math import sin, cos, asin, atan2, sqrt, radians, degrees
 import numpy as np
 from numpy.linalg import norm
 from numba import jit
-from USStandardAtmosphere import *
-from coordinate import *
+from .USStandardAtmosphere import *
+from .coordinate import *
 
 
 @jit("f8(f8,f8,f8,f8,f8)", nopython=True)
